feat: add browser support
